+++ AplikacjeMobilne-master/wsei-todo-app-master/src/screens/Home/index.tsx
@@ -1,7 +1,6 @@
 import React, { FC } from 'react';
-import { Button, View, Image, Text , StyleSheet } from 'react-native';
+import {  View, Image, Text , StyleSheet } from 'react-native';
 import styled from 'styled-components/native';
-import { useNavigation } from '@react-navigation/native';
 import Constants from 'expo-constants';
 
 import Colors from '../../constans/Colors';
@@ -19,17 +18,7 @@ const CustomImage = styled.Image`
     marginTop:10px;
     borderRadius:50px;
 `;
-const styles = StyleSheet.create({
-    container:{
-        flex:3,
-        paddingTop: Constants.statusBarHeight,
-    },
-    paragraph: {
-        margin: 15,
-        fontSize: 18,
-        textAlign: 'left',
-      },
-});
+
 
 interface IHomeProps { }
 
@@ -40,7 +29,7 @@ const Home: FC<IHomeProps> = (props) => {
             <View style={{backgroundColor:'lightgrey', alignItems:'center' , flex:1}}>
             <WelcomeText>Welcome to HomeScreen of this application</WelcomeText>
             <CustomImage
-                source={require('../../assets/YoRHa.jpg')}
+                source={require('../../assets/splash2.png')}
             />
             </View>
             <View style={{flex:1.5}}>
@@ -56,5 +45,18 @@ const Home: FC<IHomeProps> = (props) => {
         
     );
 };
+
+const styles = StyleSheet.create({
+    container:{
+        flex:3,
+        paddingTop: Constants.statusBarHeight,
+    },
+    paragraph: {
+        margin: 15,
+        fontSize: 18,
+        textAlign: 'left',
+      },
+      
+});
 
 export default Home;