+++ AplikacjeMobilne-master/wsei-todo-app-master/CHANGELOG.md
@@ -25,4 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Changed splashScreen
     - Logo changed to Circular
 
+## [1.0.5] - 2020-05-25 Added
+    - Finished TodoList screen 
+    - Minor improvements to description screen ( added more Text and additional Image)